fix: add padding to card items
This is more a temp "fix".  It adds padding to contain the image
rendered & adds a 100% height to the link container -- the purpose is to
account for images with different sizes / growth factors like svgs.
Needs a better alternative, probably rendering the image as a background
image on the card instead .. will check later